Fixed bug in fortran.y, in production ordinary_stmt where the statement is the first one, implying a main program.  The test of whether to move the %MAIN symtab entry out of scope was incorrect for the case where the statement is an INTERFACE that, unlike TYPE, does not put anything into the symbol table.
Fixed some architecture-dependent mismatches between format specs and
output items in debugging code.

Fixed bug in get_punctuation() to properly recognize OPERATOR (/) as
four tokens and not be fooled into lexing the (/ or /) as array
constructor delimiters.

Fixed bug in fortran.y, in pop_block() erroneously issuing an error
message when END INTERFACE closing an INTERFACE with generic-spec
lacks the generic-spec.  This is not an error for TYPE or INTERFACE
blocks.  I noticed that for TYPE blocks, the error was suppressed
because the block name was saved in curr_stmt_name all the way from
the opening TYPE statement to the closing END TYPE statement.  There
should be no named constructs inside a type definition block but this
is unsafe and obscure.  Also it was incorrectly ignoring the
symbolic_name on the END TYPE statement if present.  I added action
code at end_derived_type_stmt to set curr_stmt_name to NULL if not
present and to the generic-spec if present.  Then in pop_block(),
must_check_name is set to FALSE for END TYPE and END INTERFACE
statements so the error message will not be issued if the block name
is missing.

Fixed bug causing seg fault for code having a statement like PUBLIC ::
OPERATOR(+).  We have not yet implemented storing of defined operators
in the local symbol table.  The action code for this production calls
apply_attr() with the OPERATOR token as argument, which is not the
right thing to do.  This created a defective symbol table entry.  For
now, apply_attr() simply ignores any argument that is not an ID_EXPR.
Eventually we need to create proper symbol table entries for operators
and assignments as part of the semantic processing of INTERFACE blocks.

+++ ftnchek/plsymtab.c
@@ -546,8 +546,8 @@ print_arg_array(arglist)	/* prints type and flag info for arguments */
 		if(a[i].array_var)
 		    (void)fprintf(list_fd,
 			    ":%ldD(%ld)",
-			    array_dims(a[i].array_dim),
-			    array_size(a[i].array_dim)
+			    (long)(array_dims(a[i].array_dim)),
+			    (long)(array_size(a[i].array_dim))
 		    );
 		(void)fprintf(list_fd,
 			" lval%d set%d asg%d ubs%d ary%d are%d ext%d do%d",
@@ -619,8 +619,8 @@ print_com_array(cmlist)
 		    array_size(c[i].dimen_info) != 1 )
 		    (void)fprintf(list_fd,
 			    ":%ldD(%ld)",
-			    array_dims(c[i].dimen_info),
-			    array_size(c[i].dimen_info)
+			    (long)(array_dims(c[i].dimen_info)),
+				  (long)(array_size(c[i].dimen_info))
 		    );
 		(void)fprintf(list_fd,
 			" use%d set%d ubs%d asg%d",
@@ -709,16 +709,16 @@ debug_symtabs(VOID) 	/* Debugging output: hashtable and symbol tables */
 	    if(hashtab[i].name != NULL) {
 	      (void)fprintf(list_fd,"\n%4d %s",i,hashtab[i].name);
 	      if(hashtab[i].loc_symtab != NULL)
-		(void)fprintf(list_fd," loc %d",hashtab[i].loc_symtab-loc_symtab);
+		(void)fprintf(list_fd," loc %ld",(long)(hashtab[i].loc_symtab-loc_symtab));
 	      if(hashtab[i].glob_symtab != NULL)
 		(void)fprintf(list_fd,
-			" glob %d",hashtab[i].glob_symtab-glob_symtab);
+			      " glob %ld",(long)(hashtab[i].glob_symtab-glob_symtab));
 	      if(hashtab[i].com_loc_symtab != NULL)
 		(void)fprintf(list_fd,
-			" Cloc %d",hashtab[i].com_loc_symtab-loc_symtab);
+			      " Cloc %ld",(long)(hashtab[i].com_loc_symtab-loc_symtab));
 	      if(hashtab[i].com_glob_symtab != NULL)
 		(void)fprintf(list_fd,
-			" Cglob %d",hashtab[i].com_glob_symtab-glob_symtab);
+			      " Cglob %ld",(long)(hashtab[i].com_glob_symtab-glob_symtab));
 	    }
 	}
     }
